// frontend/src/core/api/generated/model/adjustmentRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export interface AdjustmentRequest {
  id: string
  materialId: string
  locationId: string
  /** @minimum 0 */
  countedQuantity: number
  /**
   * @minLength 0
   * @maxLength 120
   */
  reason: string
  day: string
  /**
   * @minLength 0
   * @maxLength 500
   * @nullable
   */
  note?: string | null
}